+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
  
 ## [Unreleased]
 
+#### Fixed
+- Fixed bug in redirect to registration in case only one VO and one group is available
+
 ## [v3.2.0]
 #### Added
 - Added filter JoinGroupsAndEduPersonEntitlement